Task3

+++ variadic_functions/3-print_all.c
@@ -23,7 +23,8 @@ if (format[i] == t_arg[j] && c)
 {
 printf(", ");
 break;
-} j++;
+}
+j++;
 }
 switch (format[i])
 {
@@ -34,7 +35,7 @@ case 'i':
 printf("%d", va_arg(valist, int)), c = 1;
 break;
 case 'f':
-printf("%d", va_arg(valist, int)), c = 1;
+printf("%f", va_arg(valist, double)), c = 1;
 break;
 case 's':
 str = va_arg(valist, char *), c = 1;
